后台登录

+++ database/migrations/2017_04_09_100358_create_permission_and_roles.php
@@ -13,32 +13,43 @@ class CreatePermissionAndRoles extends Migration
      */
     public function up()
     {
-        Schema::create("admin_roles", function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->string('description');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('admin_roles')) {
+            Schema::create("admin_roles", function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name');
+                $table->string('description');
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
 
-        Schema::create("admin_permissions", function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->string('description');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('admin_permissions')) {
+            Schema::create("admin_permissions", function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name');
+                $table->string('description');
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
 
-        Schema::create("admin_permission_role", function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer("role_id");
-            $table->integer("permission_id");
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('admin_permission_role')) {
+            Schema::create("admin_permission_role", function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer("role_id");
+                $table->integer("permission_id");
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
 
-        Schema::create("admin_role_user", function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer("role_id");
-            $table->integer("user_id");
-        });
+        if (!Schema::hasTable('admin_role_user')) {
+            Schema::create("admin_role_user", function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer("role_id");
+                $table->integer("user_id");
+            });
+        }
     }
 
     /**
@@ -48,9 +59,17 @@ class CreatePermissionAndRoles extends Migration
      */
     public function down()
     {
-        Schema::drop('admin_roles');
-        Schema::drop('admin_permissions');
-        Schema::drop('admin_permission_role');
-        Schema::drop('admin_role_user');
+        if (Schema::hasTable('admin_roles')) {
+            Schema::drop('admin_roles');
+        }
+        if (Schema::hasTable('admin_permissions')) {
+            Schema::drop('admin_permissions');
+        }
+        if (Schema::hasTable('admin_permission_role')) {
+            Schema::drop('admin_permission_role');
+        }
+        if (Schema::hasTable('admin_role_user')) {
+            Schema::drop('admin_role_user');
+        }
     }
 }
